Introduce migrations module
